Auto-configure tsConfig
