+++ example/src/routes/+page.svelte
@@ -7,4 +7,3 @@
 <MyButton text="Very cool button"/>
 <MyOtherButton deez="Hello World"/>
 <MyThirdComponent/>
-<MyOtherButton deez="Hello world"/>
