+++ example/vite.config.ts
@@ -1,12 +1,12 @@
 import { sveltekit } from '@sveltejs/kit/vite';
 import type { UserConfig } from 'vite';
-import autowire from 'sveltekit-autowire'
+import { autowire } from 'sveltekit-autowire'
 
 const config: UserConfig = {
 	plugins: [autowire({
-		components: [{ directory: './src/lib', flat: true}],
+		components: [{ directory: './src/lib', flat: true }],
 		module: {
-			"svelte" : ["onMount"]
+			"svelte": ["onMount"]
 		}
 	}), sveltekit()]
 };
