+++ plugin/src/autoimport/lib/componentMapping/createMapping.ts
